+++ futurehome/CHANGELOG.md
@@ -6,6 +6,7 @@
 - Added support for pairing new devices.
 - Added support for unpairing devices.
 - Fixed using the unit reported by the numerical sensor.
+- Added restarting the add-on when disconnected from the hub or Home Assistant
 
 ## 0.1.7 (26.07.2025)